Réarrange le code de soumission des formulaires des étapes d'homologation
