+++ public/homologation/etapes/autorite.js
@@ -1,0 +1,12 @@
+import brancheComportemenFormulaireEtape from './formulaireEtape.js';
+
+$(() => {
+  brancheComportemenFormulaireEtape((idService) => {
+    const donnees = {
+      nom: $('#nom-prenom').val(),
+      fonction: $('#fonction').val(),
+    };
+
+    return axios.put(`/api/service/${idService}/dossier/autorite`, donnees);
+  });
+});
