+++ public/homologation/etapes/dossiers.js
@@ -1,0 +1,5 @@
+import brancheComportemenFormulaireEtape from './formulaireEtape.js';
+
+$(() => {
+  brancheComportemenFormulaireEtape(() => Promise.resolve());
+});
