+++ public/homologation/etapes/etapeDate.js
@@ -1,0 +1,12 @@
+import brancheComportemenFormulaireEtape from './formulaireEtape.js';
+
+$(() => {
+  brancheComportemenFormulaireEtape((idService) => {
+    const donnees = {
+      dateHomologation: $('#date-homologation').val(),
+      dureeValidite: $('input[name="dureeValidite"]:checked').val(),
+    };
+
+    return axios.put(`/api/service/${idService}/dossier`, donnees);
+  });
+});
