+++ public/homologation/etapes/recapitulatif.js
@@ -1,0 +1,6 @@
+import brancheComportemenFormulaireEtape from './formulaireEtape.js';
+
+$(() => {
+  brancheComportemenFormulaireEtape((idService) => axios
+    .put(`/api/service/${idService}/dossier`, { finalise: true }));
+});
