+++ src/vues/service/etapeDossier/date.pug
@@ -31,6 +31,7 @@ block formulaire
         messageErreur: 'Ce champ est obligatoire. Veuillez choisir une option.',
         requis: true,
       })
+  script(type = "module", src = "/statique/homologation/etapes/etapeDate.js")
 
 block bouton-etape
   button.bouton#suivant(
