+++ src/vues/service/formulaireEtapier.pug
@@ -43,5 +43,3 @@ block zone-principale
     .boutons
       a(href = `/service/${service.id}`): .bouton.blanc Revenir à la synthèse
       block bouton-etape
-
-  script(type = 'module', src = '/statique/homologation/etapesDossier.js')
